+++ estacionamento-api/src/main/java/com/fundatec/ti20/estacionamento/service/TarifaService.java
@@ -1,10 +1,18 @@
 package com.fundatec.ti20.estacionamento.service;
 
+import com.fundatec.ti20.estacionamento.exceptions.ConflitoException;
+import com.fundatec.ti20.estacionamento.exceptions.ObjectNotFoundException;
+import com.fundatec.ti20.estacionamento.model.Conta;
 import com.fundatec.ti20.estacionamento.model.Tarifa;
+import com.fundatec.ti20.estacionamento.model.enums.TipoTarifa;
+import com.fundatec.ti20.estacionamento.model.enums.TipoVeiculo;
 import com.fundatec.ti20.estacionamento.repository.TarifaRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import java.math.BigDecimal;
+import java.util.List;
 
 
 @Service
@@ -13,32 +21,47 @@ public class TarifaService {
 
     @Autowired
     private final TarifaRepository repository;
-    @Autowired
-    private final CalcularContaService calculoService;
-    @Autowired
-    private final VeiculoService veiculoService;
 
-    public Tarifa find(Integer idVeiculo, Integer idConta) {
-        return repository.findByIdVeiculoAndIdConta(idVeiculo, idConta).orElseThrow(() -> new RuntimeException("Tarifa não encontrada"));
+
+    public Tarifa find(TipoTarifa tipoTarifa, TipoVeiculo tipoVeiculo) {
+        return repository.findByTipoTarifaAndTipoVeiculo(tipoTarifa,
+                tipoVeiculo).orElseThrow(() -> new ObjectNotFoundException("Tarifa"));
+    }
+
+    public List<Tarifa> findAll() {
+        if(repository.findAll().isEmpty()){
+            throw new ObjectNotFoundException("nenhuma tarifa, o banco de dados está vazio");
+        }
+        return repository.findAll();
+    }
+
+    public Tarifa findById(Integer id) {
+        return repository.findById(id).orElseThrow(() -> new ObjectNotFoundException("Tarifa"));
     }
 
     public Tarifa salvar(Tarifa tarifa) {
+        if (repository.findAll().contains(tarifa)){
+            throw new ConflitoException("tarifa");
+        }
         return repository.save(tarifa);
     }
 
     public void delete(Integer id) {
+        if (!repository.findAll().contains(id)){
+            throw new ObjectNotFoundException("tarifa");
+        }
         repository.deleteById(id);
     }
 
     //Revisar método requerido pelo professor
-//    public Tarifa fecharTarifa(Conta conta) {
-//
-//        double valorTarifa = calculoService.calcular(conta.getTipoVeiculo(), conta.descobrirDuracaoEmMinutos());
-//        Tarifa tarifa = Tarifa.builder()
-//                .veiculo(conta.g)
-//                .valor(new BigDecimal(valorTarifa))
-//                .build();
-//        return salvar(tarifa);
+//    public void MontarTarifa(Tarifa tarifa) {
+//        Conta contaUtilizada = tarifa.getConta();
+//        long tempoEmMinutos = contaService.descobrirDuracaoEmMinutos(contaUtilizada);
+//        TipoVeiculo tipoVeiculo = contaUtilizada.getVeiculo().getTipoVeiculo();
+//        tarifa.setValor(BigDecimal.valueOf(calculoService.calcular(tipoVeiculo, tempoEmMinutos)));
+//        contaUtilizada.setValorPagamento(tarifa.getValor());
 //    }
+
+
 
 }
